create new uuid.ts file
